Update migration and seeder for event_types and metas

+++ database/migrations/2025_08_22_180127_create_user_metas_table.php
@@ -13,19 +13,21 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('user_meta', function (Blueprint $table) {                        
-            $table->id();
-            $table->string('meta_key');
-            $table->text('meta_value')->nullable();
-            $table->unique(['user_id', 'meta_key']);
-            $table->foreignId('user_id')
-                  ->constrained('users')
-                  ->onUpdate('cascade')
-                  ->onDelete('cascade');
-            $table->index('user_id');
-            $table->index('meta_key');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('user_meta')) {
+            Schema::create('user_meta', function (Blueprint $table) {                        
+                $table->id();
+                $table->string('meta_key');
+                $table->text('meta_value')->nullable();
+                $table->unique(['user_id', 'meta_key']);
+                $table->foreignId('user_id')
+                      ->constrained('users')
+                      ->onUpdate('cascade')
+                      ->onDelete('cascade');
+                $table->index('user_id');
+                $table->index('meta_key');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
